% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGB_eq.R
\name{agb_eq}
\alias{agb_eq}
\title{Pantropical biomass equation Chave et al., 2005}
\usage{
agb_eq(vec_wd, vec_dbh, vec_height, vec_family, vec_genus)
}
\arguments{
\item{vec_wd}{Data vector with WD value (per individual) in g.cm-3}

\item{vec_dbh}{Data vector with the diameter (dbh) value in centimeters}

\item{vec_height}{Data vector with the height value in meters}

\item{vec_family}{Vector with the family name}

\item{vec_genus}{Vector with the Genus Name}
}
\value{
AGB Aboveground biomass value
}
\description{
Pantropical equation of the paper
Chave, J., Réjou-Méchain, M., Búrquez, A., Chidumayo, E., Colgan, M.S., Delitti, W.B.C., Duque, A., Eid, T., Fearnside, P.M., Goodman, R.C., Henry, M., Martínez-Yrízar, A., Mugasha, W.A., Muller-Landau, H.C., Mencuccini, M., Nelson, B.W., Ngomanda, A., Nogueira, E.M., Ortiz-Malavassi, E., Pélissier, R., Ploton, P., Ryan, C.M., Saldarriaga, J.G., Vieilledent, G., 2014. Improved allometric models to estimate the aboveground biomass of tropical trees. Glob. Change Biol. 20, 3177–3190. https://doi.org/10.1111/gcb.12629
}
\examples{
agb_eq(0.5,10,12,"Lauraceae","As")
wd<-c(0.5,0.5,0.5,0.5)
dbh<-c(10,10,10,10)
h<-c(12,12,12,12)
family<-c("Lauraceae","Arecaceae","Arecaceae","Fabaceae")
genus<-c("Ocotea","Geonoma","Euterpe","Acacia")
agb_eq(0.5,10,12,"Lauraceae","As")
agb_eq(0.5,10,12,"Arecaceae","Geonoma")
agb_eq(0.5,10,12,"Arecaceae","Euterpe")
}
