% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_eq.R
\name{height_eq}
\alias{height_eq}
\title{Height equation for the Paracou network plots}
\usage{
height_eq(vec_dbh, type_height)
}
\arguments{
\item{vec_dbh}{Tree diameter (dbh) vector in centimeters}

\item{type_height}{"Low" and "High" (There are two models, the high one is used for the P16 plot)}
}
\value{
Height value in meters
}
\description{
(Vincent et al., 2014)
Vincent, G., Sabatier, D., Rutishauser, E., 2014. Revisiting a universal airborne light detection and ranging approach for tropical forest carbon mapping: scaling-up from tree to stand to landscape. Oecologia 175, 439–443. https://doi.org/10.1007/s00442-014-2913-y
}
\examples{
height_eq(10,"low")



}
