% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_postgres.R
\name{request_postgres}
\alias{request_postgres}
\title{Query to postgres database}
\usage{
request_postgres(dbname, host, port, user, password, sql_postgres)
}
\arguments{
\item{dbname}{Database name}

\item{host}{Host name}

\item{port}{Database connection port}

\item{user}{Username}

\item{password}{password}

\item{sql_postgres}{Query database (SQL)}
}
\value{
data table with the results of the query
}
\description{
Connection to the postgres database, where inventory data is stored.
}
\examples{
datos_postgres=request_postgres(dbname="paracou",host="localhost",port=5432,user="postgres",password="postgres",sql_postgres="SELECT DISTINCT * FROM  paracou.bd_plots_corrected WHERE dbh>9.6 ORDER BY idtree, censusyear;")
head(datos_postgres,10)
}
