% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_field_data.R
\name{tree_field_data}
\alias{tree_field_data}
\title{Query to postgres database}
\usage{
tree_field_data(database)
}
\arguments{
\item{database}{Dataframe containing the database with the following columns:idtree,dbh,censusyear,dbh_dead,family,genus,wd,xfield,yfield, the two years must be included their difference is made by the attribute censusyear}
}
\value{
database with the calculated values of AGB, basal area and biovolume for two periods
}
\description{
Connection to the postgres database, where inventory data is stored.
}
